Added C# language support
Added a formatMap for the C# syntax of a Hashtable. I chose this
over the more common generic Dictionary because additional
refactoring would need to be done to account for the types and I
didn’t feel it would be worth it.

Modified the jsonToLang function to account for double quotes
in C# vs singles quotes in the existing languages implemented.